authorization on comments

+++ src/Controller/CommentController.php
@@ -40,6 +40,8 @@ class CommentController extends AbstractController
      */
     public function new(Request $request, Article $article, User $user): Response
     {
+
+
         $comment = new Comment();
         $form = $this->createForm(CommentType::class, $comment, [
             'action' => $this->generateUrl('comment_new', array('id'=>$article->getId())),
@@ -74,7 +76,7 @@ class CommentController extends AbstractController
      * @param Comment $comment
      * @return Response
      */
-    // peut-être pas très utile de laisser cette route  x
+
     public function show(Comment $comment): Response
     {
         return $this->render('comment/show.html.twig', [
@@ -84,9 +86,21 @@ class CommentController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="comment_edit", methods={"GET","POST"})
+     * @param Request $request
+     * @param Comment $comment
+     * @return Response
      */
     public function edit(Request $request, Comment $comment): Response
     {
+
+        $user = $this->getUser();
+        $author = $comment->getBlogger();
+
+        // Check if the user is the author
+        if ($user->getId() != $author->getId()) {
+            return $this->redirectToRoute("error_403");
+        }
+
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
@@ -104,9 +118,20 @@ class CommentController extends AbstractController
 
     /**
      * @Route("/{id}", name="comment_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param Comment $comment
+     * @return Response
      */
     public function delete(Request $request, Comment $comment): Response
     {
+
+        $user = $this->getUser();
+        $author = $comment->getBlogger();
+
+        // Check if the user is the author
+        if ($user->getId() != $author->getId()) {
+            return $this->redirectToRoute("error_403");
+        }
         if ($this->isCsrfTokenValid('delete' . $comment->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($comment);
